Move ILobbyGameController to common package
