+++ game-core/src/main/java/games/strategy/engine/lobby/client/ui/LobbyGameTableModel.java
@@ -15,8 +15,8 @@ import javax.swing.table.AbstractTableModel;
 import com.google.common.annotations.VisibleForTesting;
 
 import games.strategy.engine.lobby.common.ILobbyGameBroadcaster;
+import games.strategy.engine.lobby.common.ILobbyGameController;
 import games.strategy.engine.lobby.server.GameDescription;
-import games.strategy.engine.lobby.server.ILobbyGameController;
 import games.strategy.engine.message.IChannelMessenger;
 import games.strategy.engine.message.IRemoteMessenger;
 import games.strategy.engine.message.MessageContext;
